Event and movement update
Changed events to have a box collider as their effective area and added a penalty to the agent when it triggers an event when the player is not in its area. Also updated movement to use random points and a nav mesh instead of a spline for more random movement.